UIMA-2939 - use long instead of int
git-svn-id: https://svn.apache.org/repos/asf/uima/sandbox/ruta/trunk@1485764 13f79535-47bb-0310-9956-ffa450edef68

+++ ruta-core/src/main/java/org/apache/uima/ruta/RutaStream.java
@@ -690,7 +690,7 @@ public class RutaStream extends FSIteratorImplBase<AnnotationFS> {
             simpleGreedyForComposed, crowd);
   }
 
-  public int getHistogram(Type type) {
+  public long getHistogram(Type type) {
     return cas.getAnnotationIndex(type).size();
   }
